fix some files

+++ admin/delete_staff.php
@@ -1,7 +1,9 @@
 <?php
 include 'config.php';
 if (!isset($_SESSION['admin'])) { header("Location: login.php"); exit(); }
-$id=(int)$_GET['id'];
+if ($_SERVER['REQUEST_METHOD'] !== 'POST') { header("Location: view_staff.php"); exit(); }
+csrf_check();
+$id=(int)$_POST['id'];
 mysqli_query($conn,"DELETE FROM staff WHERE Staff_ID=$id");
 header("Location: view_staff.php"); exit();
 ?>
